requirements.txt
numpy
gdown
joblib==0.15.1
opencv-python
trimesh==3.8.12
pyrender==0.1.36
smplx==0.1.28
tqdm==4.47.0
loguru==0.3.2
albumentations==0.4.6
scipy==1.7.0
pyyaml==5.3.1
yacs==0.1.8
matplotlib==3.4.2
scikit-image==0.17.2
pillow==8.2.0
pytorch-lightning==1.1.8
numba==0.54.1
filterpy==1.4.5
flatten-dict==0.2.0
fastai==1.0.55
jpeg4py==0.1.4
chumpy==0.70
torchgeometry==0.1.2
setuptools==47.3.1
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
git+https://github.com/giacaglia/pytube.git
